+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
 ### This is a detailed list of all the changes made to the template.
+#### I use this since I like only making a commit or two a day, but I still want to catalog all I have done.
 
 
 #### 13/06/22, 16:50
 * Initial commit of most files
+
+#### 13/06/22, 17:33
+* Fixed typo in README.md
+* Updated CHANGELOG.md description 
+* Added logging system
+* Added the beginning of the GameWindow